Move `__memcpy_completion_impl` to its own file
