+++ libcudacxx/include/cuda/std/__cuda/barrier.h
@@ -38,6 +38,7 @@
 #include <cuda/__barrier/try_get_barrier_handle.h>
 #include <cuda/__fwd/pipeline.h>
 #include <cuda/__memcpy_async/memcpy_async_tx.h>
+#include <cuda/__memcpy_async/memcpy_completion.h>
 #include <cuda/std/__atomic/api/owned.h>
 #include <cuda/std/__type_traits/void_t.h> // _CUDA_VSTD::void_t
 
@@ -48,147 +49,6 @@
 #if defined(_CCCL_CUDA_COMPILER)
 
 _LIBCUDACXX_BEGIN_NAMESPACE_CUDA
-
-template <typename... _Ty>
-_LIBCUDACXX_HIDE_FROM_ABI constexpr bool __unused(_Ty...)
-{
-  return true;
-}
-
-template <typename _Ty>
-_LIBCUDACXX_HIDE_FROM_ABI constexpr bool __unused(_Ty&)
-{
-  return true;
-}
-
-// This struct contains functions to defer the completion of a barrier phase
-// or pipeline stage until a specific memcpy_async operation *initiated by
-// this thread* has completed.
-
-// The user is still responsible for arriving and waiting on (or otherwise
-// synchronizing with) the barrier or pipeline barrier to see the results of
-// copies from other threads participating in the synchronization object.
-struct __memcpy_completion_impl
-{
-  template <typename _Group>
-  _CCCL_NODISCARD _LIBCUDACXX_HIDE_FROM_ABI static async_contract_fulfillment
-  __defer(__completion_mechanism __cm,
-          _Group const& __group,
-          _CUDA_VSTD::size_t __size,
-          barrier<::cuda::thread_scope_block>& __barrier)
-  {
-    // In principle, this is the overload for shared memory barriers. However, a
-    // block-scope barrier may also be located in global memory. Therefore, we
-    // check if the barrier is a non-smem barrier and handle that separately.
-    if (!__is_local_smem_barrier(__barrier))
-    {
-      return __defer_non_smem_barrier(__cm, __group, __size, __barrier);
-    }
-
-    switch (__cm)
-    {
-      case __completion_mechanism::__async_group:
-        // Pre-SM80, the async_group mechanism is not available.
-        NV_IF_TARGET(
-          NV_PROVIDES_SM_80,
-          (
-            // Non-Blocking: unbalance barrier by 1, barrier will be
-            // rebalanced when all thread-local cp.async instructions
-            // have completed writing to shared memory.
-            _CUDA_VSTD::uint64_t* __bh = __try_get_barrier_handle(__barrier);
-
-            asm volatile("cp.async.mbarrier.arrive.shared.b64 [%0];" ::"r"(static_cast<_CUDA_VSTD::uint32_t>(
-              __cvta_generic_to_shared(__bh)))
-                         : "memory");));
-        return async_contract_fulfillment::async;
-      case __completion_mechanism::__async_bulk_group:
-        // This completion mechanism should not be used with a shared
-        // memory barrier. Or at least, we do not currently envision
-        // bulk group to be used with shared memory barriers.
-        _CCCL_UNREACHABLE();
-      case __completion_mechanism::__mbarrier_complete_tx:
-#  if __cccl_ptx_isa >= 800
-        // Pre-sm90, the mbarrier_complete_tx completion mechanism is not available.
-        NV_IF_TARGET(NV_PROVIDES_SM_90,
-                     (
-                       // Only perform the expect_tx operation with the leader thread
-                       if (__group.thread_rank() == 0) { ::cuda::device::barrier_expect_tx(__barrier, __size); }));
-#  endif // __cccl_ptx_isa >= 800
-        return async_contract_fulfillment::async;
-      case __completion_mechanism::__sync:
-        // sync: In this case, we do not need to do anything. The user will have
-        // to issue `bar.arrive_wait();` to see the effect of the transaction.
-        return async_contract_fulfillment::none;
-      default:
-        // Get rid of "control reaches end of non-void function":
-        _CCCL_UNREACHABLE();
-    }
-  }
-
-  template <typename _Group, thread_scope _Sco, typename _CompF>
-  _CCCL_NODISCARD _LIBCUDACXX_HIDE_FROM_ABI static async_contract_fulfillment __defer(
-    __completion_mechanism __cm, _Group const& __group, _CUDA_VSTD::size_t __size, barrier<_Sco, _CompF>& __barrier)
-  {
-    return __defer_non_smem_barrier(__cm, __group, __size, __barrier);
-  }
-
-  template <typename _Group, thread_scope _Sco, typename _CompF>
-  _LIBCUDACXX_HIDE_FROM_ABI static async_contract_fulfillment __defer_non_smem_barrier(
-    __completion_mechanism __cm, _Group const& __group, _CUDA_VSTD::size_t __size, barrier<_Sco, _CompF>& __barrier)
-  {
-    // Overload for non-smem barriers.
-
-    switch (__cm)
-    {
-      case __completion_mechanism::__async_group:
-        // Pre-SM80, the async_group mechanism is not available.
-        NV_IF_TARGET(NV_PROVIDES_SM_80,
-                     (
-                       // Blocking: wait for all thread-local cp.async instructions to have
-                       // completed writing to shared memory.
-                       asm volatile("cp.async.wait_all;" ::
-                                      : "memory");));
-        return async_contract_fulfillment::async;
-      case __completion_mechanism::__mbarrier_complete_tx:
-        // Non-smem barriers do not have an mbarrier_complete_tx mechanism..
-        _CCCL_UNREACHABLE();
-      case __completion_mechanism::__async_bulk_group:
-        // This completion mechanism is currently not expected to be used with barriers.
-        _CCCL_UNREACHABLE();
-      case __completion_mechanism::__sync:
-        // sync: In this case, we do not need to do anything.
-        return async_contract_fulfillment::none;
-      default:
-        // Get rid of "control reaches end of non-void function":
-        _CCCL_UNREACHABLE();
-    }
-  }
-
-  template <typename _Group, thread_scope _Sco>
-  _LIBCUDACXX_HIDE_FROM_ABI static async_contract_fulfillment
-  __defer(__completion_mechanism __cm, _Group const& __group, _CUDA_VSTD::size_t __size, pipeline<_Sco>& __pipeline)
-  {
-    // pipeline does not sync on memcpy_async, defeat pipeline purpose otherwise
-    __unused(__pipeline);
-    __unused(__size);
-    __unused(__group);
-
-    switch (__cm)
-    {
-      case __completion_mechanism::__async_group:
-        return async_contract_fulfillment::async;
-      case __completion_mechanism::__async_bulk_group:
-        return async_contract_fulfillment::async;
-      case __completion_mechanism::__mbarrier_complete_tx:
-        return async_contract_fulfillment::async;
-      case __completion_mechanism::__sync:
-        return async_contract_fulfillment::none;
-      default:
-        // Get rid of "control reaches end of non-void function":
-        _CCCL_UNREACHABLE();
-    }
-  }
-};
 
 /***********************************************************************
  * memcpy_async code:
